[#1] Hande  Month dropdown onChange
