+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@gooddata/react-components": "6.2.0",
     "http-proxy-middleware": "0.19.1",
+    "moment": "^2.26.0",
     "node-sass": "4.11.0",
     "react": "16.5.2",
     "react-dom": "16.5.2",
