+++ src/components/grossProfitInMonth/grossProfitInMonth.css
@@ -1,0 +1,3 @@
+div.GrossProfitInMonth {
+    height: 460px;
+}
