+++ src/index.css
@@ -4,10 +4,6 @@ body {
     font-family: sans-serif;
 }
 
-.App > div {
-    height: 400px;
-}
-
 .App select {
     font-size: 30px;
 }
